+++ test/io_utils_test.dart
@@ -17,4 +17,31 @@ void main() {
       expect(getFileNameFor('dartdoc.generator'), 'dartdoc-generator.html');
     });
   });
+
+  group('printWarningDelta', () {
+    Map<String, int> original;
+    Map<String, int> current;
+    setUp(() {
+      original = new Map.fromIterables(["originalwarning", "morewarning", "duplicateoriginalwarning"],
+                                       [1, 1, 2]);
+      current = new Map.fromIterables(["newwarning", "morewarning", "duplicateoriginalwarning"],
+          [1, 1, 1]);
+    });
+
+    test('verify output of printWarningDelta', () {
+      expect(printWarningDelta('Diff Title', 'oldbranch', original, current),
+          equals('*** Diff Title : 1 warnings from original (oldbranch) missing in current:\n'
+              'originalwarning\n'
+              '*** Diff Title : 1 new warnings not in original (oldbranch)\n'
+              'newwarning\n'
+              '*** Diff Title : Identical warning quantity changed\n'
+              '* Appeared 2 times in original (oldbranch), now 1:\n'
+              'duplicateoriginalwarning\n'));
+    });
+
+    test('verify output when nothing changes', () {
+      expect(printWarningDelta('Diff Title 2', 'oldbranch2', original, original),
+          equals('*** Diff Title 2 : No difference in warning output from original (oldbranch2) (3 warnings found)\n'));
+    });
+  });
 }
